Cleanup!
* Halo and augmented star stencils have to wait for v1.x.
* So does dynamic repartitioning.
* Split the mesh stencil tests into star and halo stencil tests.
* Removed game of life tests for now.
* Fixed the parallel star stencil unit tests.
