+++ model/mesh_stencils.h
@@ -16,12 +16,14 @@
 // cell from one of its neighboring cells. This unweighted stencil is 
 // constructed for every cell in the given mesh, and does not include the 
 // "central" cell.
+// NOTE: This stencil is not currently implemented for radius > 1!
 stencil_t* cell_star_stencil_new(mesh_t* mesh, int radius);
 
 // Creates a halo-shaped stencil for the cells in the given mesh. The halo 
 // consists of a single layer of cells surrounding the nodes of each cell.
 // This unweighted stencil is constructed for every cell in the given mesh, 
 // and does not include the "central" cell.
+// NOTE: This stencil is not currently implemented!
 stencil_t* cell_halo_stencil_new(mesh_t* mesh);
 
 #endif
