+++ model/tests/test_halo_stencil.c
@@ -15,67 +15,6 @@
 #include "geometry/create_uniform_mesh.h"
 #include "model/mesh_stencils.h"
 
-void test_NXxNYxNZ_star_stencil(void** state, 
-                                MPI_Comm comm,
-                                int radius,
-                                int nx, int ny, int nz,
-                                int num_interior_neighbors, 
-                                int num_boundary_neighbors, 
-                                int num_edge_neighbors,
-                                int num_corner_neighbors)
-{
-  bbox_t bbox = {.x1 = 0.0, .x2 = 1.0, .y1 = 0.0, .y2 = 1.0, .z1 = 0.0, .z2 = 1.0};
-  mesh_t* mesh = create_uniform_mesh(comm, nx, ny, nz, &bbox);
-  cubic_lattice_t* lattice = cubic_lattice_new(nx, ny, nz);
-  stencil_t* stencil = cell_star_stencil_new(mesh, radius);
-  for (int i = 0; i < nx; ++i)
-  {
-    for (int j = 0; j < ny; ++j)
-    {
-      for (int k = 0; k < nz; ++k)
-      {
-        // The cell score is a way of telling whether we're in the 
-        // interior (0), on a boundary but not an edge/corner (1), on an 
-        // edge but not a corner (2), or on a corner (3).
-        int cell_score = 0;
-        if ((i == 0) || (i == nx-1))
-          ++cell_score;
-        if ((j == 0) || (j == ny-1))
-          ++cell_score;
-        if ((k == 0) || (k == nz-1))
-          ++cell_score;
-
-        // Count the number of neighbors.
-        int cell = cubic_lattice_cell(lattice, i, j, k);
-        int num_neighbors = stencil_size(stencil, cell);
-        if (cell_score == 0)
-          assert_int_equal(num_interior_neighbors, num_neighbors);
-        else if (cell_score == 1)
-          assert_int_equal(num_boundary_neighbors, num_neighbors);
-        else if (cell_score == 2)
-          assert_int_equal(num_edge_neighbors, num_neighbors);
-        else if (cell_score == 3)
-          assert_int_equal(num_corner_neighbors, num_neighbors);
-
-        // Make sure the neighbors are all valid, distinct cell indices.
-        int_unordered_set_t* neighbors = int_unordered_set_new();
-        int pos = 0, n;
-        real_t weight;
-        while (stencil_next(stencil, cell, &pos, &n, &weight))
-        {
-          assert_true(n >= 0);
-          assert_true(n < mesh->num_cells);
-          assert_false(int_unordered_set_contains(neighbors, n));
-          int_unordered_set_insert(neighbors, n);
-        }
-        int_unordered_set_free(neighbors);
-      }
-    }
-  }
-  stencil_free(stencil);
-  mesh_free(mesh);
-}
-
 void test_NXxNYxNZ_halo_stencil(void** state, 
                                 MPI_Comm comm,
                                 int nx, int ny, int nz,
@@ -136,26 +75,6 @@ void test_NXxNYxNZ_halo_stencil(void** state,
   mesh_free(mesh);
 }
 
-void test_serial_1x1x1_cell_star_stencil(void** state)
-{
-  test_NXxNYxNZ_star_stencil(state, MPI_COMM_SELF, 1, 1, 1, 1, 0, 0, 0, 0);
-}
-
-void test_serial_10x1x1_cell_star_stencil(void** state)
-{
-  test_NXxNYxNZ_star_stencil(state, MPI_COMM_SELF, 1, 10, 1, 1, 0, 0, 2, 1);
-}
-
-void test_serial_10x10x1_cell_star_stencil(void** state)
-{
-  test_NXxNYxNZ_star_stencil(state, MPI_COMM_SELF, 1, 10, 10, 1, 4, 4, 3, 2);
-}
-
-void test_serial_10x10x10_cell_star_stencil(void** state)
-{
-  test_NXxNYxNZ_star_stencil(state, MPI_COMM_SELF, 1, 10, 10, 10, 6, 5, 4, 3);
-}
-
 void test_serial_1x1x1_cell_halo_stencil(void** state)
 {
   test_NXxNYxNZ_halo_stencil(state, MPI_COMM_SELF, 1, 1, 1, 0, 0, 0, 0);
@@ -174,31 +93,6 @@ void test_serial_10x10x1_cell_halo_stencil(void** state)
 void test_serial_10x10x10_cell_halo_stencil(void** state)
 {
   test_NXxNYxNZ_halo_stencil(state, MPI_COMM_SELF, 10, 10, 10, 26, 17, 11, 7);
-}
-
-void test_parallel_1x1x1_cell_star_stencil(void** state)
-{
-  test_NXxNYxNZ_star_stencil(state, MPI_COMM_WORLD, 1, 1, 1, 1, 0, 0, 0, 0);
-}
-
-void test_parallel_10x1x1_cell_star_stencil(void** state)
-{
-  test_NXxNYxNZ_star_stencil(state, MPI_COMM_WORLD, 1, 10, 1, 1, 0, 0, 2, 1);
-}
-
-void test_parallel_10x10x1_cell_star_stencil(void** state)
-{
-  test_NXxNYxNZ_star_stencil(state, MPI_COMM_WORLD, 1, 10, 10, 1, 4, 4, 3, 2);
-}
-
-void test_parallel_10x10x10_cell_star_stencil(void** state)
-{
-  test_NXxNYxNZ_star_stencil(state, MPI_COMM_WORLD, 1, 10, 10, 10, 6, 5, 4, 3);
-}
-
-void test_parallel_1x1x1_cell_halo_stencil(void** state)
-{
-  test_NXxNYxNZ_halo_stencil(state, MPI_COMM_WORLD, 1, 1, 1, 0, 0, 0, 0);
 }
 
 void test_parallel_10x1x1_cell_halo_stencil(void** state)
@@ -221,21 +115,12 @@ int main(int argc, char* argv[])
   polymec_init(argc, argv);
   const UnitTest tests[] = 
   {
-    unit_test(test_serial_1x1x1_cell_star_stencil),
-    unit_test(test_serial_10x1x1_cell_star_stencil),
-    unit_test(test_serial_10x10x1_cell_star_stencil),
-    unit_test(test_serial_10x10x10_cell_star_stencil),
     unit_test(test_serial_1x1x1_cell_halo_stencil),
     unit_test(test_serial_10x1x1_cell_halo_stencil),
     unit_test(test_serial_10x10x1_cell_halo_stencil),
     unit_test(test_serial_10x10x10_cell_halo_stencil)
 #if POLYMEC_HAVE_MPI
-   ,unit_test(test_parallel_1x1x1_cell_star_stencil),
-    unit_test(test_parallel_10x1x1_cell_star_stencil),
-    unit_test(test_parallel_10x10x1_cell_star_stencil),
-    unit_test(test_parallel_10x10x10_cell_star_stencil),
-    unit_test(test_parallel_1x1x1_cell_halo_stencil),
-    unit_test(test_parallel_10x1x1_cell_halo_stencil),
+   ,unit_test(test_parallel_10x1x1_cell_halo_stencil),
     unit_test(test_parallel_10x10x1_cell_halo_stencil),
     unit_test(test_parallel_10x10x10_cell_halo_stencil)
 #endif
